Extract duplicated AddHeaderSigner to separate source file
